Adapt Linux releasing
Provide script configuration options. By default, Linux configuration
is intended to be used when installing the game on Linux.

Linux version of the game is designed to be installed by compiling
from source. Remove creation of Linux release package, which was
built on Ubuntu CI host. Executable of release package required a
different version of glibc than what was present on Debian virtual
machine which was used for testing.

The options for Linux releasing are:
1. Selecting a set of supported distributions and implementing a
   build for each.
2. Advocating building from source.

Advocating building from source is initially selected. The option
of building is present anyway and the process is straight forward
with the provided install scripts. Build instructions still need
to be provided and there is an issue in issue tracker for this.