Remove libsdl package install from runtime dependency install script
In addition to SDL (possibly from libsdl package), SDL_image is
needed. There is Debian package libsdl-image1.2, but no any
libsdl-image2.

It is anyway a scetchy idea to build with one version of libraries
and run with another version of libraries. Proceeding to recommend
building SDL and SDL_image from source when installing them as
runtime dependencies.